backend/requirements.txt
# --- Core UI ---
gradio>=4.0.0

# --- Gemini (NEW SDK, not deprecated) ---
google-genai>=0.3.0

# --- Resume parsing ---
pdfplumber>=0.11.0
pymupdf>=1.23.0

# --- Job search & HTTP ---
requests>=2.31.0

# --- Voice output ---
gTTS>=2.5.0

# --- Utilities ---
python-dotenv>=1.0.



fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.10.3
